% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{CobbDouglas}
\alias{CobbDouglas}
\title{Single Output Data Generation}
\usage{
CobbDouglas(N, nX)
}
\arguments{
\item{N}{Sample size.}

\item{nX}{Number of inputs. Possible values: \code{1}, \code{2}, \code{3},
\code{4}, \code{5},\code{6}, \code{9}, \code{12} and \code{15}.}
}
\value{
\code{data.frame} with the simulated data.
}
\description{
This function is used to simulate the data in a single output
scenario.
}
