% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smooth.R
\name{CreateCubicBF}
\alias{CreateCubicBF}
\title{Generate a new pair of Cubic Basis Functions}
\usage{
CreateCubicBF(data, xi, knt, B, side)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in
the model.}

\item{xi}{Variable index of the new basis function(s).}

\item{knt}{Knots for creating the new basis function(s).}

\item{B}{Matrix of basis functions.}

\item{side}{Side of the basis function.}
}
\value{
Matrix of basis functions updated with the new basis functions.
}
\description{
This function generates two new cubic basis functions from a
variable and a knot previously created during MARS algorithm.
}
