% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{DEA}
\alias{DEA}
\alias{DEA_object}
\title{LS-Boosting with adapted Multivariate Adaptive Frontier Splines (MARS)}
\usage{
DEA(data, x, y)

DEA_object(data, x, y, pred, score)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{pred}{Output predictions using the BBC output measure}

\item{score}{Efficiency score using the BBC output measure}
}
\value{
A \code{DEA} object.

A \code{DEA} object.
}
\description{
This function estimates a production frontier satisfying some classical production theory axioms, such as monotonicity and concavity, which is based upon the adaptation of the machine learning technique known as LS-boosting using adapted Multivariate Adaptive Regression Splines (MARS) as base learners.

This function saves information about the DEA model.
}
