% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{EAT_object}
\alias{EAT_object}
\title{Create a EAT object}
\usage{
EAT_object(data, x, y, rownames, numStop, max.leaves, na.rm, tree)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{rownames}{\code{string}. Data rownames.}

\item{numStop}{Minimum number of observations in a node for a split to be attempted.}

\item{max.leaves}{Depth of the tree.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted. If \code{FALSE}, an error occurs in case of \code{NA} rows.}

\item{tree}{\code{list} containing the nodes of the Efficiency Analysis Trees pruned model.}
}
\value{
An \code{EAT} object.
}
\description{
This function saves information about the Efficiency Analysis Trees model.
}
