% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EATBoost.R
\name{EATBoost}
\alias{EATBoost}
\title{Gradient Tree Boosting}
\usage{
EATBoost(data, x, y, num.iterations, num.leaves, learning.rate)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{num.iterations}{Maximum number of iterations the algorithm will perform}

\item{num.leaves}{Maximum number of reflected pairs created by the forward algorithm of MARS.}

\item{learning.rate}{Learning rate that control the overfitting of the algorithm. Value must be in (0,1]}
}
\value{
A \code{EATBoost} object.
}
\description{
This function estimates a production frontier satisfying some classical production theory axioms, such as monotonicity and concavity, which is based upon the adaptation of the machine learning technique known as LS-boosting using adapted Multivariate Adaptive Regression Splines (MARS) as base learners.
}
