% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForwardAlgorithm.R
\name{EstimCoeffsForward}
\alias{EstimCoeffsForward}
\title{Estimate Coefficients in Multivariate Adaptive Frontier Splines during Forward Procedure.}
\usage{
EstimCoeffsForward(B, y)
}
\arguments{
\item{B}{\code{matrix} of basis functions.}

\item{y}{Output \code{vector} in data.}
}
\value{
\code{vector} with the coefficients estimated.
}
\description{
This function solves a Quadratic Programming Problem to obtain a set of coefficients.
}
