% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDH.R
\name{FDH}
\alias{FDH}
\alias{FDH_object}
\title{Free Disposal Hull model}
\usage{
FDH(data, x, y)

FDH_object(data, x, y, pred, score)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in
the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{pred}{Output predictions using the BBC radial output measure}

\item{score}{Efficiency score using the BBC radial output measure}
}
\value{
A \code{FDH} object.

A \code{FDH} object.
}
\description{
This function estimates a production frontier satisfying Free
Disposal HUll axioms using the radial output measure.

This function saves information about the FDH model.
}
