% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{get.a.EATBoost}
\alias{get.a.EATBoost}
\title{Get \code{EATBoost} leaves supports}
\usage{
get.a.EATBoost(EATBoost_model)
}
\arguments{
\item{EATBoost_model}{Model from class \code{EATBoost} from which the data
are obtained}
}
\value{
\code{data.frame} with the leave supports
}
\description{
Calculates the inferior corner of the leaves supports of a
\code{EATBoost} model.
}
