% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{get.b.trees}
\alias{get.b.trees}
\title{Get the superior corner of the leave support from all trees
of \code{EATBoost}}
\usage{
get.b.trees(EATBoost_model)
}
\arguments{
\item{EATBoost_model}{Model from class \code{EATBoost} from which the data
are obtained}
}
\value{
\code{list} of \code{matrix}. The length of the list is equal to
the \code{num.iterations} of the \code{EATBoost_model}. Each \code{matrix}
corresponds to a tree,  where the number of columns is the number of input
 variables and the number of rows to the number of leaves
}
\description{
Calculates the superior corner of the support of all leave nodes
of every tree created in the \code{EATBoost} model
}
