% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{get.intersection.a}
\alias{get.intersection.a}
\title{Get intersection between two leaves supports}
\usage{
get.intersection.a(current_a, current_b, new_a, new_b)
}
\arguments{
\item{current_a}{Inferior corner of first leave support}

\item{current_b}{Superior corner of first leave support}

\item{new_a}{Inferior corner of second leave support}

\item{new_b}{Superior corner of second leave support}
}
\value{
\code{vector} with the intersection. \code{NULL} if intersection
is not valid.
}
\description{
Calculates the intersection between two leave nodes from
different trees of a \code{EATBoost} model.
}
