% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForwardAlgorithm.R
\name{mse}
\alias{mse}
\title{Mean Squared Error}
\usage{
mse(y, yPred)
}
\arguments{
\item{y}{Vector of actual data.}

\item{yPred}{Vector of predicted values.}
}
\value{
Mean Squared Error.
}
\description{
This function computes the mean squared error between two
numeric vectors.
}
