% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\name{predict.DEA}
\alias{predict.DEA}
\title{Model Prediction for DEA}
\usage{
\method{predict}{DEA}(object, newdata, x, measure = "rad.out")
}
\arguments{
\item{object}{A \code{MARSBoost} object.}

\item{newdata}{\code{data.frame}. Set of input variables to predict on.}

\item{x}{Inputs index.}

\item{measure}{Model for prediction. \code{1} MARS Boost without smoothing.\code{2} MARS Boost with smoothing.}
}
\value{
\code{data.frame} with the predicted values.
}
\description{
This function predicts the expected output by a \code{MARSBoost} object.
}
