% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\name{predict.EAT}
\alias{predict.EAT}
\title{Model Prediction for Efficiency Analysis Trees.}
\usage{
\method{predict}{EAT}(object, newdata, x, ...)
}
\arguments{
\item{object}{An \code{EAT} object.}

\item{newdata}{\code{data.frame}. Set of input variables to predict on.}

\item{x}{Inputs index.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{data.frame} with the predicted values.
}
\description{
This function predicts the expected output by an \code{EAT} object.
}
