% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\name{predict.EATBoost}
\alias{predict.EATBoost}
\title{Model prediction for EATBoost algorithm}
\usage{
\method{predict}{EATBoost}(object, newdata, x, ...)
}
\arguments{
\item{object}{A \code{EATBoost} object.}

\item{newdata}{\code{data.frame}. Set of input variables to predict on.}

\item{x}{Inputs index.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{data.frame} with the predicted values.
}
\description{
This function predicts the expected output by a \code{EATBoost}
object.
}
