% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\name{predict.MARSBoost}
\alias{predict.MARSBoost}
\title{Model Prediction for Boosted Multivariate Adaptive Frontier Splines}
\usage{
\method{predict}{MARSBoost}(object, newdata, x, class = 1, ...)
}
\arguments{
\item{object}{A \code{MARSBoost} object.}

\item{newdata}{\code{data.frame}. Set of input variables to predict on.}

\item{x}{Inputs index.}

\item{class}{Model for prediction. \code{1} MARS Boost without smoothing.
\code{2} MARS Boost with smoothing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{data.frame} with the predicted values.
}
\description{
This function predicts the expected output by a \code{MARSBoost}
object.
}
