% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{predictDEA}
\alias{predictDEA}
\title{Model prediction for DEA}
\usage{
predictDEA(data, x, y)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}
}
\value{
\code{data.frame} with the original data and the predicted values through a DEA model.
}
\description{
This function predicts the expected output through a DEA model.
}
