% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDH.R
\name{predictFDH_BBC_out}
\alias{predictFDH_BBC_out}
\title{Model prediction for FDH}
\usage{
predictFDH_BBC_out(
  data,
  x,
  y,
  dataOriginal = data,
  xOriginal = x,
  yOriginal = y
)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the new variables
in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{dataOriginal}{\code{data.frame} or \code{matrix} containing the
original variables used to create the model.}

\item{xOriginal}{Vector. Column input indexes in original data.}

\item{yOriginal}{Vector. Column output indexes in original data.}
}
\value{
\code{data.frame} with the the predicted values through a FDH model
and the efficiency score
}
\description{
This function predicts the expected output through a FDH model.
}
