% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{preProcess}
\alias{preProcess}
\title{Data Pre-processing for Multivariate Adaptive Frontier Splines.}
\usage{
preProcess(data, x, y, na.rm = TRUE)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{na.rm}{\code{logical} If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
It returns a \code{data.frame} in the required format.
}
\description{
This function arranges the data in the required format and displays error messages.
}
