% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{get.intersection.a}
\alias{get.intersection.a}
\title{Get intersection between two leaves supports}
\usage{
get.intersection.a(comb_a_actual, comb_b_actual)
}
\arguments{
\item{comb_a_actual}{Inferior corner of first leave support}

\item{comb_b_actual}{Superior corner of first leave support}
}
\value{
\code{vector} with the intersection. \code{NULL} if intersection
is not valid.
}
\description{
Calculates the intersection between two leave nodes from
different trees of a \code{EATBoost} model.
}
