% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\name{predict.DEA}
\alias{predict.DEA}
\title{Model Prediction for DEA}
\usage{
\method{predict}{DEA}(object, newdata, x, y, ...)
}
\arguments{
\item{object}{A \code{DEA} object.}

\item{newdata}{\code{data.frame}. Set of input variables to predict on.}

\item{x}{Inputs index.}

\item{y}{Outputs index.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{data.frame} with the predicted values. Valid measures are:
\code{rad.out}.
}
\description{
This function predicts the expected output by a \code{DEA}
object.
}
